backend/requirements.txt
fastapi
uvicorn[standard]
httpx
google-generativeai
supabase
python-dotenv
pydantic
